requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
dnspython==1.16.0
docopt==0.6.2
eventlet==0.24.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
monotonic==1.5
multidict==4.5.2
pipreqs==0.4.9
psycopg2==2.8.2
python-engineio==3.5.1
python-socketio==4.0.1
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
websocket-client==0.56.0
whitenoise==4.1.2
yarg==0.1.9
yarl==1.3.0
